requirements_kaggle.txt
torch==2.0.1
transformers==4.30.2
sentencepiece==0.1.99
datasets==2.13.0
py-fire==0.0.3
spacy==3.5.0
tqdm==4.65.0
numpy==1.24.3
